refactor : 폴더 구조 수정 및 에러 코드 정의
